Extract style entity & Add class decorator
