+++ packages/engine/src/base/AbstractObserver.ts
@@ -9,6 +9,26 @@ export function Property () {
     }
 }
 
+export function Properties (includes: string[] = []) {
+    return (target: any) => {
+        const keys = Object.keys(target.prototype)
+        if (!target.prototype.hasOwnProperty('_properties')) {
+            target.prototype._properties = []
+        }
+        for (const i of keys) {
+            if (target.prototype.hasOwnProperty(i)) {
+                const number: number = includes.findIndex((value: string) => {
+                    return i === value
+                })
+                if (number === -1 && typeof target[i] !== 'function') {
+                    target.prototype._properties.push(i)
+                }
+            }
+        }
+    }
+
+}
+
 abstract class AbstractObserver implements Observer {
     // noinspection JSMismatchedCollectionQueryUpdate
     private readonly _properties?: string[]
@@ -16,35 +36,35 @@ abstract class AbstractObserver implements Observer {
     protected constructor () {
     }
 
-     public resolve (): void {
+    public resolve (): void {
         if (this._properties === undefined) {
             return
         }
         for (const key of this._properties) {
-             if (this.hasOwnProperty(key)) {
-                 const field = Object.getOwnPropertyDescriptor(this, key)
-                 if (field === undefined) {
-                     continue
-                 }
-                 if (!field.configurable) {
-                     continue
-                 }
-                 const property = key.replace('_', '')
-                 Object.defineProperty(this, property, {
-                     get () {
-                         return this[key]
-                     },
-                     set (v) {
-                         this.notify()
-                         this[key] = v
-                     }
-                 })
-             }
-         }
+            if (this.hasOwnProperty(key)) {
+                const field = Object.getOwnPropertyDescriptor(this, key)
+                if (field === undefined) {
+                    continue
+                }
+                if (!field.configurable) {
+                    continue
+                }
+                const property = key.replace('_', '')
+                Object.defineProperty(this, property, {
+                    get () {
+                        return this[key]
+                    },
+                    set (v) {
+                        this.notify()
+                        this[key] = v
+                    }
+                })
+            }
+        }
     }
 
     public initialize (): void {
-         this.resolve()
+        this.resolve()
     }
 
     public abstract notify (): void
